+++ Makefile
@@ -4,8 +4,8 @@ CFLAGS = -Wall -Wextra -I./headers
 LDFLAGS = 
 
 # Define object files for each program to improve clarity and reduce redundancy
-SERVER_OBJS = ./bin/server.o ./bin/server_connection.o ./bin/message.o ./bin/checksum.o ./bin/request_send_file.o
-CLIENT_OBJS = ./bin/client.o ./bin/client_connection.o ./bin/message.o ./bin/checksum.o ./bin/request_send_file.o
+SERVER_OBJS = ./bin/server.o ./bin/server_connection.o ./bin/message.o ./bin/checksum.o
+CLIENT_OBJS = ./bin/client.o ./bin/client_connection.o ./bin/message.o ./bin/checksum.o 
 TEST_OBJS = ./bin/checksum_test.o ./bin/checksum.o ./bin/message.o # Group test objects for clarity
 
 # ---
@@ -17,6 +17,16 @@ all: build
 
 # 'build' target relies on 'mkdir_bin' and then builds the executables
 build: mkdir_bin server client
+# 'build' target relies on 'mkdir_bin' and then builds the executables
+build: mkdir_bin server client
+
+# New rule to create the bin directory - idempotent
+mkdir_bin:
+	@mkdir -p ./bin
+
+# ---
+# Executable Targets
+# ---
 
 # New rule to create the bin directory - idempotent
 mkdir_bin:
@@ -28,9 +38,11 @@ mkdir_bin:
 
 server: $(SERVER_OBJS)
 	@echo "Linking server executable"
+	@echo "Linking server executable"
 	$(CC) $(CFLAGS) $(SERVER_OBJS) -o ./bin/server $(LDFLAGS)
 
 client: $(CLIENT_OBJS)
+	@echo "Linking client executable"
 	@echo "Linking client executable"
 	$(CC) $(CFLAGS) $(CLIENT_OBJS) -o ./bin/client $(LDFLAGS)
 
@@ -65,8 +77,8 @@ test: checksum_test
 ./bin/checksum.o: ./cores/common/checksum.c | mkdir_bin
 	$(CC) $(CFLAGS) -c $< -o $@
 
-./bin/request_send_file.o: ./cores/common/request_send_file.c | mkdir_bin
-	$(CC) $(CFLAGS) -c $< -o $@
+# ./bin/request_send_file.o: ./cores/common/request_send_file.c | mkdir_bin
+# 	$(CC) $(CFLAGS) -c $< -o $@
 
 ./bin/checksum_test.o: ./tests/checksum_test.c | mkdir_bin
 	$(CC) $(CFLAGS) -c $< -o $@
@@ -76,14 +88,23 @@ test: checksum_test
 # ---
 
 # Clears all contents of the bin directory
+# ---
+# Cleanup Targets
+# ---
+
+# Clears all contents of the bin directory
 clear:
+	@echo "Clearing ./bin directory"
 	@echo "Clearing ./bin directory"
 	@rm -rf ./bin/*
 
 # 'clean' usually removes all build artifacts, often synonym for 'clear'
+# 'clean' usually removes all build artifacts, often synonym for 'clear'
 clean: clear
 
 # Target to specifically remove only object files, if needed
+# Target to specifically remove only object files, if needed
 clear-o:
+	@echo "Clearing object files"
 	@echo "Clearing object files"
 	@rm -f ./bin/*.o
