+++ cores/common/request_send_file.c
@@ -1,127 +1,127 @@
-#include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
-#include <sys/socket.h>
-#include <arpa/inet.h>
-#include <unistd.h>
-#include "connection.h"
-#include "message.h"
-#include "checksum.h"
+// #include <stdio.h>
+// #include <string.h>
+// #include <stdlib.h>
+// #include <sys/socket.h>
+// #include <arpa/inet.h>
+// #include <unistd.h>
+// #include "connection.h"
+// #include "message.h"
+// #include "checksum.h"
 
 
-void segment_file(const char* filename, int sock, struct sockaddr_in client, struct sockaddr_in server) { 
-    char buffer[BUFFER_SIZE];
-    FILE *file = fopen(filename, "rb");
-    if (!file) {
-        perror("Failed to open file");
-        return;
-    }
+// void segment_file(const char* filename, int sock, struct sockaddr_in client, struct sockaddr_in server) { 
+//     char buffer[BUFFER_SIZE];
+//     FILE *file = fopen(filename, "rb");
+//     if (!file) {
+//         perror("Failed to open file");
+//         return;
+//     }
 
-    struct message msg;
-    memset(&msg, 0, sizeof(msg));
+//     struct message msg;
+//     memset(&msg, 0, sizeof(msg));
     
-    // Send metadata packet
-    HDR_SET_META(msg.flags);
-    strcpy(msg.data, filename);
-    msg.data_length = strlen(msg.data);
-    set_message_checksum(&msg);
-    send_message(&msg, sock, client);
+//     // Send metadata packet
+//     HDR_SET_META(msg.flags);
+//     strcpy(msg.data, filename);
+//     msg.data_length = strlen(msg.data);
+//     set_message_checksum(&msg);
+//     send_message(&msg, sock, client);
 
-    size_t bytes_read;
-    uint32_t seqNum = 0x0;
+//     size_t bytes_read;
+//     uint32_t seqNum = 0x0;
 
-    while ((bytes_read = fread(buffer, 1, BUFFER_SIZE, file)) > 0) {
-        memset(&msg, 0, sizeof(msg));
-        HDR_SET_SEQ(msg.flags, seqNum);
-        memcpy(msg.data, buffer, bytes_read);
-        msg.data_length = bytes_read;
-        set_message_checksum(&msg);
-        send_message(&msg, sock, client);
-        seqNum++;
-    }
+//     while ((bytes_read = fread(buffer, 1, BUFFER_SIZE, file)) > 0) {
+//         memset(&msg, 0, sizeof(msg));
+//         HDR_SET_SEQ(msg.flags, seqNum);
+//         memcpy(msg.data, buffer, bytes_read);
+//         msg.data_length = bytes_read;
+//         set_message_checksum(&msg);
+//         send_message(&msg, sock, client);
+//         seqNum++;
+//     }
 
-    // Send final packet with FIN flag
-    memset(&msg, 0, sizeof(msg));
-    HDR_SET_FIN(msg.flags);
-    msg.data_length = 0;
-    set_message_checksum(&msg);
-    send_message(&msg, sock, client);
+//     // Send final packet with FIN flag
+//     memset(&msg, 0, sizeof(msg));
+//     HDR_SET_FIN(msg.flags);
+//     msg.data_length = 0;
+//     set_message_checksum(&msg);
+//     send_message(&msg, sock, client);
 
-    fclose(file);
-    printf("File %s segmented and sent.\n", filename);
-}
+//     fclose(file);
+//     printf("File %s segmented and sent.\n", filename);
+// }
 
-int request_file(char fileName[], int sock, struct sockaddr_in server, struct sockaddr_in client_addr)
-{
-    struct message msg;
-    memset(&msg, 0, sizeof(msg));
+// int request_file(char fileName[], int sock, struct sockaddr_in server, struct sockaddr_in client_addr)
+// {
+//     struct message msg;
+//     memset(&msg, 0, sizeof(msg));
 
-    strncpy(msg.data, fileName, sizeof(msg.data));
-    msg.data_length = strlen(msg.data);
-    HDR_SET_ACK(msg.flags, HDR_ACK_NONE);
+//     strncpy(msg.data, fileName, sizeof(msg.data));
+//     msg.data_length = strlen(msg.data);
+//     HDR_SET_ACK(msg.flags, HDR_ACK_NONE);
 
-    set_message_checksum(&msg);
+//     set_message_checksum(&msg);
     
-    printf("REQUEST: %s\n", msg.data);
-    send_message(&msg, sock, server);
+//     printf("REQUEST: %s\n", msg.data);
+//     send_message(&msg, sock, server);
 
-    struct message received_msg;
-    FILE *file = NULL;
+//     struct message received_msg;
+//     FILE *file = NULL;
 
-    while (1)
-    {
-        memset(&received_msg, 0, sizeof(received_msg));
-        socklen_t server_len = sizeof(server);
-        int len = recvfrom(sock, &received_msg, sizeof(received_msg), 0, (struct sockaddr *)&server, &server_len);
+//     while (1)
+//     {
+//         memset(&received_msg, 0, sizeof(received_msg));
+//         socklen_t server_len = sizeof(server);
+//         int len = recvfrom(sock, &received_msg, sizeof(received_msg), 0, (struct sockaddr *)&server, &server_len);
         
-        if (len < 0) {
-            perror("recvfrom failed");
-            continue;
-        }
+//         if (len < 0) {
+//             perror("recvfrom failed");
+//             continue;
+//         }
 
-        received_msg.data_length = len - (sizeof(received_msg.checksum) + sizeof(received_msg.flags));
+//         received_msg.data_length = len - (sizeof(received_msg.checksum) + sizeof(received_msg.flags));
 
-        if (validate_message_checksum(&received_msg, len) != 0) {
-            fprintf(stderr, "Checksum validation failed! Packet dropped.\n");
-            continue;
-        }
+//         if (validate_message_checksum(&received_msg, len) != 0) {
+//             fprintf(stderr, "Checksum validation failed! Packet dropped.\n");
+//             continue;
+//         }
 
-        if (HDR_GET_META(received_msg.flags)) {
-            printf("RECEIVE META DATA\n");
-            file = fopen(received_msg.data, "wb");
-            if (!file) {
-                perror("Failed to open file for writing");
-                return 1;
-            }
-        }
-        else {
-            printf("RECEIVE DATA\n");
-            fwrite(received_msg.data, 1, received_msg.data_length, file);
-        }
+//         if (HDR_GET_META(received_msg.flags)) {
+//             printf("RECEIVE META DATA\n");
+//             file = fopen(received_msg.data, "wb");
+//             if (!file) {
+//                 perror("Failed to open file for writing");
+//                 return 1;
+//             }
+//         }
+//         else {
+//             printf("RECEIVE DATA\n");
+//             fwrite(received_msg.data, 1, received_msg.data_length, file);
+//         }
 
-        if (HDR_GET_FIN(received_msg.flags)) {
-            printf("END\n");
+//         if (HDR_GET_FIN(received_msg.flags)) {
+//             printf("END\n");
             
-            // Send final ACK after receiving FIN
-            memset(&msg, 0, sizeof(msg));
-            HDR_SET_ACK(msg.flags, HDR_ACK_ACK);
-            msg.data_length = 0;
-            set_message_checksum(&msg);
-            send_message(&msg, sock, server);
+//             // Send final ACK after receiving FIN
+//             memset(&msg, 0, sizeof(msg));
+//             HDR_SET_ACK(msg.flags, HDR_ACK_ACK);
+//             msg.data_length = 0;
+//             set_message_checksum(&msg);
+//             send_message(&msg, sock, server);
             
-            break;
-        }
+//             break;
+//         }
 
-        // Send ACK for data packet
-        memset(&msg, 0, sizeof(msg));
-        HDR_SET_ACK(msg.flags, HDR_ACK_ACK);
-        msg.data_length = 0;
-        set_message_checksum(&msg);
-        send_message(&msg, sock, server);
-    }
+//         // Send ACK for data packet
+//         memset(&msg, 0, sizeof(msg));
+//         HDR_SET_ACK(msg.flags, HDR_ACK_ACK);
+//         msg.data_length = 0;
+//         set_message_checksum(&msg);
+//         send_message(&msg, sock, server);
+//     }
     
-    if (file) fclose(file);
-    printf("Successfully received the file from the server\n");
-    printf("File %s reassembled successfully.\n", fileName);
-    return 0;
-}
+//     if (file) fclose(file);
+//     printf("Successfully received the file from the server\n");
+//     printf("File %s reassembled successfully.\n", fileName);
+//     return 0;
+// }
